Add "\[" and "\]" to the response retrieved from the function summarise_response
This enables mod_jazzquiz (and other activity modules) question to render student responses with MathJax.
It avoids the need of actually adding the prefix "\[" and postfix "\]" to the student answer, stored in the db.
As a result, all the standard question functions like specifying correct answers, auto grading etc. works as normal.